Corrections FareCalculatorService

+++ parkingsystem-master/src/main/java/com/parkit/parkingsystem/service/FareCalculatorService.java
@@ -9,18 +9,18 @@ import com.parkit.parkingsystem.model.Ticket;
 public class FareCalculatorService {
     /**
      * The method calculates the fare for the given ticket
+     *
      * @param ticket stores the entry data for a vehicle
      */
- public void calculateFare(Ticket ticket) {
-     
+    public void calculateFare(Ticket ticket) {
+        
         if ((ticket.getOutTime() == null) || (ticket.getOutTime().before(ticket.getInTime()))) {
             throw new IllegalArgumentException("Out time provided is incorrect:" + ticket.getOutTime().toString());
         }
-        //TODO: Some tests are failing here. Need to check if this logic is correct
-        //Correction below:
-        long inHour = ticket.getInTime().getTime();
-        long outHour = ticket.getOutTime().getTime();
-        long duration = (outHour - inHour) / 1000 / 60;
+        
+        double inHour = ticket.getInTime().getTime();
+        double outHour = ticket.getOutTime().getTime();
+        double duration = (outHour - inHour) / 1000 / 60;
         if (duration <= 30) {
             ticket.setPrice(0);
         } else {
@@ -37,8 +37,8 @@ public class FareCalculatorService {
                     throw new IllegalArgumentException("Unkown Parking Type");
             }
             
-            if (ticket.getRecurringVehicle()==true) {
-                ticket.setPrice(ticket.getPrice()*0.95);
+            if (ticket.getRecurringVehicle() == true) {
+                ticket.setPrice(ticket.getPrice() * 0.95);
             }
         }
     }
